+++ src/components/textfield/TextField.js
@@ -17,4 +17,8 @@ export class TextFieldComponent extends BaseComponent {
     info.changeEvent = 'input';
     return info;
   }
+
+  get emptyValue() {
+    return '';
+  }
 }